+++ CHANGELOG.md
@@ -10,3 +10,4 @@
   - icon size 添加 x3s 选项
 - [InputItem](./components/input-item/index.tsx)
   - 把 InputItem clear icon 设置为 x3s
+  - 增加 `itemStyle?: ViewStyle` 属性